gd3: Initial skeleton of gluster daemon version three
This is an initial proof of concept about how glusterd could be
implemented in terms of the mgmt resource model using libmgmt.

We aim to prove that this can provide powerful automation for GlusterFS,
and also to demonstrate that mgmt can be a useful foundation for your
existing distributed software project.

A quick note on versions! Glusterd 1.x is the original C version of this
code base and the de-facto standard as of this writing. It is what is
actually used in production and provides basic management facilities.

Glusterd 2.x is a multi year project to rewrite Glusterd 1.x in golang
and to possibly include advanced management features into the project.
As of this writing the project has still not showed much success, and
attempts to integrate mgmt into that code base have proved difficult.

Glusterd 3.x (gd3) is this project. By using mgmt to replace many of the
facilities which have been partially implemented in Glusterd 2.x and by
discarding all of the unnecessary bloat and bad designs, we aim to
produce a lean management engine for GlusterFS. We'll see how it goes.

Learn more about mgmt here: https://github.com/purpleidea/mgmt/